requirements.txt
apache-airflow==2.5.0
pandas==1.5.3
scikit-learn==1.2.2
numpy==1.24.2
joblib==1.2.0
python-dotenv==1.0.0
boto3==1.26.135  # Для интеграции с AWS S3
google-api-python-client==2.86.0  # Для интеграции с Google Drive